Handle subscriptions in background jobs.
